+++ controllers/flight.controller.js
@@ -1,10 +1,6 @@
 const pool = require('../utils/dbConfig')
-/**
- * @TODO
- * -- do some better error handling when origin or destination is not found 
- *    in method fetchOriginCoordinates or fetchDestinationCoordinates 's else block
- * -- formatNumberWithCommas doesn't work
- */
+const os = require('os');
+const hostname = os.hostname();
 const {
     EARTH_RADIUS,
     FARE_PER_KM_AIR_INDIA,
@@ -13,8 +9,6 @@ const {
     FARE_PER_KM_VISTARA,
 } = require('../constants/numbers')
 
-const { CURRENCY_SYMBOL } = require('../constants/strings')
-
 async function getFlightFare(req, res) {
 
     const {
@@ -22,7 +16,7 @@ async function getFlightFare(req, res) {
         fareIndigo,
         fareSpiceJet,
         fareVistara
-    } = await calculateFlightFare(req.params.origin, req.params.destination)
+    } = await calculateFlightFare(req.params.origin, req.params.destination, res)
 
     const fares = {
         fareAIrIndia,
@@ -31,78 +25,69 @@ async function getFlightFare(req, res) {
         fareVistara
     };
 
-    // Round off the numbers to 2 decimal places and format with commas
+    // Round off the numbers to 2 decimal places
     const roundedFares = {};
     for (const key in fares) {
         const roundedValue = fares[key].toFixed(2);
-        
-        const formattedValue = `${formatNumberWithCommas(roundedValue)}`;
-        roundedFares[key] = formattedValue;
+        roundedFares[key] = roundedValue;
     }
-    return res.status(200).json({INR:roundedFares})
+
+    return res.status(200).json({ INR: roundedFares })
 }
 
-async function calculateFlightFare(origin, destination) {
+async function calculateFlightFare(origin, destination, res) {
+    try {
+        const { lat: originLat, lon: originLon } = await fetchCoordinates(origin)
+        const { lat: destLat, lon: destLon } = await fetchCoordinates(destination)
 
-    const { originLat, originLon } = await fetchOriginCoordinates(origin)
-    const { destLat, destLon } = await fetchDestinationCoordinates(destination)
+        const originLatRad = toRadians(originLat)
+        const originLonRad = toRadians(originLon)
+        const destLatRad = toRadians(destLat)
+        const destLonRad = toRadians(destLon)
 
-    const originLatRad = toRadians(originLat)
-    const originLonRad = toRadians(originLon)
-    const destLatRad = toRadians(destLat)
-    const destLonRad = toRadians(destLon)
+        // Calculate the differences between the latitudes and longitudes
+        const latDiff = destLatRad - originLatRad
+        const lngDiff = destLonRad - originLonRad
 
-    // Calculate the differences between the latitudes and longitudes
-    const latDiff = destLatRad - originLatRad
-    const lngDiff = destLonRad - originLonRad
+        // calculating the distance between two coordinates using the HAVERSINE formula
+        const a = Math.sin(latDiff / 2) ** 2 + Math.cos(originLatRad) * Math.cos(destLatRad) * Math.sin(lngDiff / 2) ** 2
+        const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a))
+        const distance = EARTH_RADIUS * c
 
-    // calculatting the distance bet two coordinates using the Haversine formula
-    const a = Math.sin(latDiff / 2) ** 2 + Math.cos(originLatRad) * Math.cos(destLatRad) * Math.sin(lngDiff / 2) ** 2
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a))
-    const distance = EARTH_RADIUS * c
+        const fareAIrIndia = distance * FARE_PER_KM_AIR_INDIA
+        const fareIndigo = distance * FARE_PER_KM_INDIGO
+        const fareSpiceJet = distance * FARE_PER_KM_SPICEJET
+        const fareVistara = distance * FARE_PER_KM_VISTARA
 
-    const fareAIrIndia = distance * FARE_PER_KM_AIR_INDIA
-    const fareIndigo = distance * FARE_PER_KM_INDIGO
-    const fareSpiceJet = distance * FARE_PER_KM_SPICEJET
-    const fareVistara = distance * FARE_PER_KM_VISTARA
-
-    return {
-        fareAIrIndia,
-        fareIndigo,
-        fareSpiceJet,
-        fareVistara
+        return {
+            fareAIrIndia,
+            fareIndigo,
+            fareSpiceJet,
+            fareVistara
+        }
+    } catch (error) {
+        console.log('Error in calculating flight fare:', error);
+        return res.status(500).json({ error: error.message })
     }
 }
 
-async function fetchOriginCoordinates(origin) {
-    const query = `SELECT * FROM airport_data WHERE icao = '${origin}'`
+// console.log('Current hostname:', hostname);
+async function fetchCoordinates(icao_code) {
+    const query = `SELECT * FROM airport_data WHERE icao = '${icao_code}'`
     const result = await pool.query(query)
 
     if (result.rowCount === 1) {
-        const originLat = result.rows[0].lat
-        const originLon = result.rows[0].lon
-        return { originLat, originLon }
+        const lat = result.rows[0].lat
+        const lon = result.rows[0].lon
+        return { lat, lon }
     } else {
-        return { error: 'Origin not found' };
-    }
-}
-
-async function fetchDestinationCoordinates(destination) {
-    const query = `SELECT * FROM airport_data WHERE icao = '${destination}'`
-    const result = await pool.query(query)
-    if (result.rowCount === 1) {
-        const destLat = result.rows[0].lat
-        const destLon = result.rows[0].lon
-        return { destLat, destLon }
-    } else {
-        return { error: 'Destination not found' };
+        const error = `couldn\'t find the airport, make sure you are using the correct ICAO code, Please refer to ${hostname}/airports/airports.json`
+        throw new Error(error);
     }
 }
 
 function toRadians(degrees) {
     return degrees * (Math.PI / 180)
 }
-function formatNumberWithCommas(number) {
-    return number.toLocaleString('en-IN');
-}
+
 module.exports = { getFlightFare }